Remove -isEqual: and -hash from OTPToken
These methods were untested and unused. Originally, I intended them to be used when checking if a given token was already in the token manager's array. It turns out that the only property that truly mattered in that case was the keychainItemRef, which -isEqual: wasn't even considering. It's actually okay if the array contains two "equal" tokens, as long as they are truly two different items in the keychain.