Restore full token deletion animation
Every token list view model update – even those without any row changes – was triggering a table view update. As a result, the row deletion animation when a token was deleted was cut short by the (empty) animation set of the next update. The fix is to prevent table updates except when truly necessary.