Prototype of using privacy screen and LocalAuthentication
- Adds Component that:
  - Checks if device can use LocalAuthentication
  - Checks if a successful auth challenge has occurred

Given that LocalAuthentication is available on the device:

When the application becomes foreground after launching a privacy screen is presented. A successful
LocalAuthentication dismisses the privacy screen.

When the application enters the background state the privacy screen is presented. This prevents
tokens from being displayed during app switching.

None of the keychain items are using LocalAuthentication for encryption. This is purely UI related
so the security/encryption of the keychain items have not been changed by this feature.

Tokens are still readable/displayable by the app no matter what the state of the LocalAuthentication
challenge is.